--- pickee_frontend/src/components/TestComponent.vue
@@ -1,21 +1,0 @@
-<template>
-    <v-combobox
-        v-model="select"
-        :items="items"
-        label="Combobox"
-        multiple
-        chips
-    ></v-combobox>
-</template>
-
-<script>
-    export default {
-        name: "TestComponent",
-        data: () => ({
-            select: ["Vuetify", "Programming"],
-            items: ["Programming", "Design", "Vue", "Vuetify"]
-        })
-    };
-</script>
-
-<style scoped></style>
