--- pickee_frontend/src/mainTwo.js
@@ -1,8 +1,0 @@
-import Vue from 'vue'
-import App from './App02.vue'
-
-Vue.config.productionTip = false
-
-new Vue({
-  render: h => h(App),
-}).$mount('#app')
