+++ pickee_frontend/src/pages/About/About.vue
@@ -1,0 +1,24 @@
+<template>
+    <div id="about">
+        <v-app>
+            <v-content>
+                <p>Home</p>
+            </v-content>
+        </v-app>
+    </div>
+</template>
+
+<script>
+    export default {
+        name: "About",
+        data: () => ({}),
+        methods: {},
+        components: {}
+    };
+</script>
+
+<style>
+    #about {
+        margin: 20px;
+    }
+</style>
