+++ pickee_frontend/src/pages/Home/Home.vue
@@ -1,0 +1,24 @@
+<template>
+    <div id="home">
+        <v-app>
+            <v-content>
+                <p>Home</p>
+            </v-content>
+        </v-app>
+    </div>
+</template>
+
+<script>
+    export default {
+        name: "Home",
+        data: () => ({}),
+        methods: {},
+        components: {}
+    };
+</script>
+
+<style>
+    #home {
+        margin: 20px;
+    }
+</style>
