+++ pickee_frontend/src/pages/Login/Login.vue
@@ -1,0 +1,24 @@
+<template>
+    <div id="login">
+        <v-app>
+            <v-content>
+                <p>Login</p>
+            </v-content>
+        </v-app>
+    </div>
+</template>
+
+<script>
+    export default {
+        name: "Login",
+        data: () => ({}),
+        methods: {},
+        components: {}
+    };
+</script>
+
+<style>
+    #login {
+        margin: 20px;
+    }
+</style>
