+++ pickee_frontend/src/pages/ProfileHistory/ProfileHistory.js
@@ -1,0 +1,14 @@
+import Vue from "vue";
+import ProfileHistory from "./ProfileHistory.vue";
+
+import vuetify from "@/plugins/vuetify/vuetify";
+
+import "@/assets/styles/app.css";
+// import "@/assets/styles/pages/profile-history.css";
+
+Vue.config.productionTip = false;
+
+new Vue({
+    vuetify,
+    render: h => h(ProfileHistory)
+}).$mount("#history");
