+++ pickee_frontend/src/pages/SignUp/SignUp.vue
@@ -1,0 +1,24 @@
+<template>
+    <div id="signup">
+        <v-app>
+            <v-content>
+                <p>Home</p>
+            </v-content>
+        </v-app>
+    </div>
+</template>
+
+<script>
+    export default {
+        name: "signup",
+        data: () => ({}),
+        methods: {},
+        components: {}
+    };
+</script>
+
+<style>
+    #signup {
+        margin: 20px;
+    }
+</style>
